+++ NEWS.md
@@ -1,11 +1,12 @@
-# amp.sim 0.2.0
+# amp.dm 0.2.0
 
 - Categorized the pkgdown reference section
 - General optimization for testing
 - Fixed issues for `attr_factor` to retain attributes and added option to set largest group first (#11)
-- Added function `flag_outliers` to flag observations for outliers based on box.plot stats 
+- Added function `flag_outliers` to flag observations for outliers based on box.plot stats
+- Added function `check_cat` to chaeck categories in a dataset
 
-# amp.sim 0.1.0
+# amp.dm 0.1.0
 
 This the initial version of the packge, evolved from mutliple inhouse developed functions and packages